requirements.txt
streamlit
ultralytics
PyYAML
numpy
opencv-python-headless
Pillow
torch
streamlit_image_zoom
# torch ***BE CAREFUL, >2.2 CUDA SUPPORT BROKEN ON WINDOWS, PYTHON 3.12, CUDA 12.4***USE DOWNLOAD FOR 11.8 FROM PYTORCH START LOCALLY WEBSITE***
# torchvision
# torchaudio
